+++ package.json
@@ -13,6 +13,7 @@
     "@chakra-ui/react": "^2.5.1",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
+    "@hookform/resolvers": "^3.0.1",
     "@reduxjs/toolkit": "^1.9.3",
     "axios": "^1.3.4",
     "framer-motion": "^10.0.1",
@@ -21,6 +22,7 @@
     "react-calendar": "^4.0.0",
     "react-datepicker": "^4.10.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.9",
     "react-icons": "^4.7.1",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.2",
@@ -30,7 +32,8 @@
     "sass": "^1.58.3",
     "swiper": "^9.1.0",
     "swr": "^2.1.0",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.0",
+    "yup": "^1.0.2"
   },
   "devDependencies": {
     "@types/react": "^18.0.27",
