+++ src/pages/administration/components/doctors/components/form/index.js
@@ -1,0 +1,2 @@
+export * from './Form.doctors.administration';
+export * from './hooks';
