+++ src/pages/administration/components/doctors/components/form/schema/index.js
@@ -1,0 +1,1 @@
+export * from './createDoctor.schema';
