+++ src/pages/administration/components/doctors/components/index.js
@@ -1,0 +1,1 @@
+export * from './form/Form.doctors.administration';
